Using EloquentTree in another database
I am working in a project with multiple databases. And we use the Eloquent Tree at one database who at not the application database default. This line code causes an error because the application it was looking the table in wrong database.

Please forgive me the english. I'm still practing.